backend/requirements.txt
cohere==5.5.3
qdrant-client==1.9.1
requests==2.31.0
beautifulsoup4==4.12.2
PyYAML==6.0.1
python-dotenv==1.0.0
pytest==7.4.3